+++ eloquentjs/11_1_tracking_the_scalpel/availableNeighbors.js
@@ -1,0 +1,9 @@
+function availableNeighbors(nest) {
+    let requests = nest.connections.map(neighbor => {
+            return request(nest, neighbor, "ping")
+            .then(() => true, () => false);
+    });
+    return Promise.all(requests).then(result => {
+            return nest.connections.filter((_, i) => result[i]);
+    });
+}
