--- eloquentjs/11_1_tracking_the_scalpel/requestTypes.js
@@ -1,24 +1,0 @@
-
-let requestTypes = {
-
-}
-
-function defineRequestType(requestTypeName, callback) {
-    requestTypes[requestTypeName] = callback;
-}
-
-function requestType(name, handler) {
-    defineRequestType(name, (nest, content, source, callback) => {
-        try {
-            Promise.resolve(handler(nest, content, source))
-            .then(response => callback(null, response), 
-            failure => callback(failure));
-        } catch (exception) {
-            callback(exception);
-        }
-    });
-}
-
-//defineRequestType("note", () => console.log("note delivered"));
-
-module.exports = {requestType, requestTypes};
